feat(site): improved toc.
I noticed that with long one-word titles, the toc view becomes ugly.
Minus sign in one line and the word in another. I decided to allow to break
the title in any character using the 'word-break: break-all' CSS property.
The fix is not elegant, but it works as I expected.